+++ src/Infrastructure/LeanCode.Serialization/JsonSnakeCaseNamingPolicy.cs
@@ -1,0 +1,9 @@
+using System.Text.Json;
+
+namespace LeanCode.Serialization
+{
+    public class JsonSnakeCaseNamingPolicy : JsonNamingPolicy
+    {
+        public override string ConvertName(string name) => name.ToSnakeCase();
+    }
+}