+++ benchmarks/LeanCode.Benchmarks/InProcCQRS.Queries.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Autofac;
 using Autofac.Extensions.DependencyInjection;
 using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
 using LeanCode.Cache.AspNet;
 using LeanCode.Components;
 using LeanCode.CQRS.Cache;
@@ -13,7 +14,7 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace LeanCode.Benchmarks
 {
-    [CoreJob]
+    [SimpleJob(RuntimeMoniker.NetCoreApp31)]
     [MarkdownExporterAttribute.Atlassian]
     [MemoryDiagnoser]
     public class InProcCQRS__Queries