+++ src/Example/CQRS/SendNotification.cs
@@ -8,6 +8,6 @@ namespace LeanCode.Example.CQRS
     public class SendNotification : IRemoteCommand
     {
         public Guid UserId { get; set; }
-        public string Content { get; set; }
+        public string Content { get; set; } = null!;
     }
 }